Updated.
git-svn-id: https://svn.r-project.org/R-dev-web/trunk@5632 c52295ea-58df-0310-926a-d16021944841

+++ CRAN/QA/Kurt/lib/bash/check_R_cp_logs.sh
@@ -47,56 +47,30 @@ for flavor in ${R_flavors}; do
     ## everything else.
     (head -1 ${d}/00check.log | grep -q '^\* using log directory') || continue
     package=`basename ${d} .Rcheck`
-    cp ${d}/00check.log ${target_dir}/${flavor}/${package}-00check.txt
-    ## If installation failed, all the check log says is
-    ##   Installation failed.
-    ##   See '/var/mnt/.....' for details.
-    ## which is not too helpful.  Provide the install log as well in
-    ## such a case, and massage the check log accordingly.
-    ## <NOTE>
-    ## We used to grep -E for
-    ##   '^\*+ checking whether.*can be installed \.\.\. ERROR$'
-    ##   '^\*+ checking whether.*can be installed \.\.\. WARNING$'
-    ## but that does not work for CRLF 00check.log files as produced by
-    ## Windows.  We could preprocess using sed 's/\r//', but let's keep
-    ## things simple and add possible space before the end-of-string.
-    ## </NOTE>
-    url="${target_url}/${flavor}/${package}-00install.html"
-    msg="See '${url}' for details."
-    if grep -E \
-        '^\*+ checking whether.*can be installed \.\.\. *(\[.*\])? *ERROR[[:space:]]*$' \
-        ${d}/00check.log > /dev/null; then
-      (head -n -1 ${d}/00check.log; echo "${msg}") \
-	> ${target_dir}/${flavor}/${package}-00check.txt
+    if test -f ${d}/00install.out; then
       cp ${d}/00install.out \
-        ${target_dir}/${flavor}/${package}-00install.txt
-    ## Also provide the install log and try pointing to it in case there
-    ## were installation warnings (only works for R 2.5.0 or better).
-    elif grep -E \
-        '^\*+ checking whether.*can be installed \.\.\. *(\[.*\])? *WARNING[[:space:]]*' \
-        ${d}/00check.log > /dev/null; then
-      sed "s|^See ['‘].*Rcheck/00install.out['’] for details.|${msg}|" \
-        ${d}/00check.log > \
-        ${target_dir}/${flavor}/${package}-00check.txt
-      cp ${d}/00install.out \
-        ${target_dir}/${flavor}/${package}-00install.txt
-    ## If running the tests failed, provide the last 13 lines of the
-    ## offending test out file.  Only works for 1.8 or better, as this
-    ## moves .Rout to .Rout.fail on error, and provided that the tests
-    ## results are available, which currently is no longer true when
-    ## collecting the results for different flavors on the cran server.
-    ## But from 2.4.0, this is also no longer necessary, as R CMD check
-    ## now record the same information in the check log.
-    elif grep -E '^\*+ checking tests \.\.\. ERROR[[:space:]]*$' \
-        ${d}/00check.log > /dev/null; then
-      bad_test_out=`ls ${d}/tests/*.fail 2>/dev/null | head -1`
-      if test -n "${bad_test_out}"; then
-        bad_test_src=`basename ${bad_test_out} out.fail`
-        (echo "Running tests/${bad_test_src} failed.";
-	  echo "Last 13 lines of output:"; echo;
-	  tail -13 ${bad_test_out}) >> \
-	  ${target_dir}/${flavor}/${package}-00check.txt
+         ${target_dir}/${flavor}/${package}-00install.txt
+      ## We used to grep -E for
+      ##   '^\*+ checking whether.*can be installed \.\.\. ERROR$'
+      ##   '^\*+ checking whether.*can be installed \.\.\. WARNING$'
+      ## but that does not work for CRLF 00check.log files as produced by
+      ## Windows.  We could preprocess using sed 's/\r//', but let's keep
+      ## things simple and add possible space before the end-of-string.
+      url="${target_url}/${flavor}/${package}-00install.html"
+      msg="See '${url}' for details."
+      if grep -E '^\*+ checking whether.*can be installed \.\.\. *(\[.*\])? *(ERROR|WARNING)[[:space:]]*$' \
+	      ${d}/00check.log > /dev/null; then
+	sed "s|^See ['‘].*Rcheck/00install.out['’] for details.|${msg}|" \
+            ${d}/00check.log > \
+            ${target_dir}/${flavor}/${package}-00check.txt
+      else
+	sed -E "s@(^\*+ checking whether.*can be installed \.\.\. *(\[.*\])? *OK[[:space:]]*$)@\1\n$msg@" \
+            ${d}/00check.log > \
+            ${target_dir}/${flavor}/${package}-00check.txt
       fi
+    else
+      cp ${d}/00check.log \
+	 ${target_dir}/${flavor}/${package}-00check.txt
     fi
   done
   ## <FIXME>
